+++ kittens/diff/main.go
@@ -112,7 +112,6 @@ func main(_ *cli.Command, opts_ *Options, args []string) (rc int, err error) {
 		return 1, err
 	}
 	init_caches()
-	create_formatters()
 	defer func() {
 		for tdir := range remote_dirs {
 			os.RemoveAll(tdir)
@@ -154,6 +153,7 @@ func main(_ *cli.Command, opts_ *Options, args []string) (rc int, err error) {
 		if !tc.KeyboardProtocol {
 			return fmt.Errorf("This terminal does not support the kitty keyboard protocol, or you are running inside a terminal multiplexer that is blocking querying for kitty keyboard protocol support. The diff kitten cannot function without it.")
 		}
+		h.on_capabilities_received(tc)
 		return nil
 	}
 	lp.OnWakeup = h.on_wakeup